+++ src/functions/learning_rate.rs
@@ -8,7 +8,7 @@ use crate::core::*;
 /// * `lr_type` - Learning rate type
 /// * `iteration` - Iteration
 /// * `som` - SOM Object
-pub fn calc_learning_rate(lr_type: &LearningRateType, iteration: u32, som: &SOM) -> f32 {
+pub fn calc_learning_rate(lr_type: &LearningRateType, iteration: usize, som: &SOM) -> f32 {
   return match lr_type {
     LearningRateType::Constant      => constant_learning_rate(som.learning_rate),
     LearningRateType::Linear        => linear_learning_rate(som.learning_rate, iteration),
@@ -23,16 +23,16 @@ fn constant_learning_rate(starting_learning_rate: f32) -> f32 {
 }
 
 /// Linear Learning Rate
-fn linear_learning_rate(starting_learning_rate: f32, iteration: u32) -> f32 {
+fn linear_learning_rate(starting_learning_rate: f32, iteration: usize) -> f32 {
   return starting_learning_rate * (1 / iteration) as f32;
 }
 
 /// Inverse of Time Learning Rate
-fn inverse_of_time_learning_rate(starting_learning_rate: f32, iteration: u32, iteration_limit: u32) -> f32 {
+fn inverse_of_time_learning_rate(starting_learning_rate: f32, iteration: usize, iteration_limit: usize) -> f32 {
   return starting_learning_rate * (1 - iteration / iteration_limit) as f32;
 }
 
 /// Power Series Learning Rate
-fn power_series_learning_rate(starting_learning_rate: f32, iteration: u32, iteration_limit: u32) -> f32 {
+fn power_series_learning_rate(starting_learning_rate: f32, iteration: usize, iteration_limit: usize) -> f32 {
   return starting_learning_rate * ((iteration / iteration_limit) as f32).exp();
 }
